Backend/requirement.txt
Flask
flask-cors
PyPDF2
python-docx